+++ Assets/Script/Resources/Enemy/EnemySO.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class EnemySO : ScriptableObject
 {
    public string enemyName = "Enemy";
-   public int hpMax = 2;
-   public int dame = 1;
+   public float hpMax = 2;
+   public float dame = 1;
    public List<DropRate> dropList;
 }
